deleteBoard 수정

+++ src/boards/boards.service.ts
@@ -49,8 +49,8 @@ export class BoardsService {
     }
 
     // 특정게시글 지우기
-    async deleteBoard(id: number): Promise <void> {
-        const result = await this.boardRepository.delete(id);
+    async deleteBoard(id: number, user: User): Promise <void> {
+        const result = await this.boardRepository.delete({id, userId: user.id});
         
         if(result.affected === 0){
             throw new NotFoundException(`Can't find Board with id ${id}`);
